overloaded the CalcDistance Function to allo
Overloaded the CalcDistance function so it can take Creatures and PCs as inputs instead of just CombatLocations. This way the function can handle Large and Huge creatures and still calculate the right distances to the nearest edges of creatures. 

So scripts in dsPCAttack and dsCreatureAttack scripts need to be updated when they check for opportunity attacks etc.