Spell that summons a prop in the middle of combat
If you make a spell that has a "Point" target you can use this spell in combat to summon the prop "Campfire Off"! Always useful to add a campfire to a combat (joke). It's more for illustrative purposes to show how to add a prop in the middle of an encounter.