Update Convo.cs
There was a major bug where you could create a conversation the first time and it would save properly. If you ever changed that conversation in the editor, it wouldsnt save over the old existing xml file. It would create a file with no ending file type and your edits would not show up the next time you opened up the toolset. This seems to fix all of that.